Rename ToDoList to TaskList and add event name to save file function
